Moving mob_index_data.h out of merc.h
